refactor views.py into utils.py
