+++ form_builder/utils.py
@@ -2,7 +2,115 @@
 from django.contrib import messages
 from .models import Fastq
 import yaml
-from .views import *
+from django.http import HttpResponse
+
+def make_CLIP_form(form, request):
+    '''
+    If download CLIPPER button is clicked, generates list of FASTQs, 
+    and yaml file is downloaded onto the page. If adding sample, creates a new 
+    FASTQ object. Otherwise, deletes FASTQ object.
+    '''
+    if request.POST.get("save"):        # "Save and Download (CLIPper)" button
+        fastqs = []
+        for key in request.POST.keys():  # generate list of fastqs
+            try:
+                if key.startswith('fqid_'):
+                    Fastq.objects.get(pk=request.POST.get(key))
+                    fastqs.append(request.POST.get(key))
+            except Exception as e:
+                print(e)
+                pass
+
+        if form.is_valid():             # save valid form and download YAML
+            clip = form.save(commit=False)
+            clip.fastqs = ','.join(fastqs)
+            clip.save()
+            file_data = CLIP_yaml(clip)
+            response = HttpResponse(
+                file_data, content_type='application/text charset=utf-8')
+            f = request.POST.get('dataset', 'manifest')
+            response['Content-Disposition'] = f'attachment; filename="{f}.yaml"'
+            return response
+
+    elif request.POST.get("newItem"):   # "Add Sample" button
+        ip_fastq_path = request.POST.get("ip_fastq_path")
+        ip_adapter_path = request.POST.get("ip_adapter_path")
+        sminput_fastq_path = request.POST.get("sminput_fastq_path")
+        sminput_adapter_path = request.POST.get("sminput_adapter_path")
+        cells = request.POST.get("cells")
+        experiment = request.POST.get("experiment")
+        sample = request.POST.get("sample")
+        ip_rep = request.POST.get("ip_rep")
+        sminput_rep = request.POST.get("sminput_rep")
+        submitter = request.user
+
+        make_fastq(ip_fastq_path, ip_adapter_path, ip_rep, sminput_fastq_path,
+                   sminput_adapter_path, sminput_rep,
+                   cells, experiment, sample, submitter, request)
+    else:                               # Delete fastq
+        for key in request.POST.keys():
+            if key.startswith('delete_fqid_'):
+                Fastq.objects.filter(
+                    id=int(key.split('delete_fqid_')[1])).delete()
+
+def make_SKIPPER_form(form, request):
+    '''
+    If download manifest button is clicked, generates list of FASTQs, 
+    and csv file is downloaded onto the page. If download config button is 
+    clicked, will download a 'Skipper_config.py file'. If adding sample, 
+    creates a new FASTQ object. Otherwise, deletes FASTQ object.
+    '''
+    if request.POST.get("save_manifest"):   # "Save and Download MANIFEST"
+        fastqs = []
+        for key in request.POST.keys():     # generate list of fastqs
+            try:
+                if key.startswith('fqid_'): 
+                    Fastq.objects.get(pk=request.POST.get(key))
+                    fastqs.append(request.POST.get(key))
+            except Exception as e:
+                print(e)  
+                pass
+
+        if form.is_valid():
+            clip = form.save(commit=False)
+            clip.fastqs = ','.join(fastqs)
+            clip.save()
+            file_data = skipper_tsv(clip)   # generate tsv 
+            response = HttpResponse(
+                file_data, content_type='application/text charset=utf-8')
+            f = request.POST.get('manifest', 'manifest.csv')
+            response['Content-Disposition'] = f'attachment; filename="{f}"'
+            return response
+    elif request.POST.get("save_config"):   # "Save and Download CONFIG"
+        if form.is_valid():
+            clip = form.save(commit=False)
+            clip.save()
+            file_data = skipper_config(clip)
+            response = HttpResponse(
+                file_data, content_type='application/text charset=utf-8')
+            f = 'Skipper_config.py'
+            response['Content-Disposition'] = f'attachment; filename="{f}"'
+            return response
+    elif request.POST.get("newItem"):       # "Add Sample" button
+        ip_fastq_path = request.POST.get("ip_fastq_path", None)
+        ip_adapter_path = request.POST.get("ip_adapter_path", None)
+        sminput_fastq_path = request.POST.get("sminput_fastq_path", None)
+        sminput_adapter_path = request.POST.get("sminput_adapter_path", None)
+        cells = request.POST.get("cells", None)
+        experiment = request.POST.get("experiment", None)
+        sample = request.POST.get("sample", None)
+        ip_rep = request.POST.get("ip_rep", None)
+        sminput_rep = request.POST.get("sminput_rep", None)
+        submitter = request.user
+
+        make_fastq(ip_fastq_path, ip_adapter_path, ip_rep, sminput_fastq_path,
+                    sminput_adapter_path, sminput_rep,
+                    cells, experiment, sample, submitter, request)
+    else:                                   # Delete fastq
+        for key in request.POST.keys():
+            if key.startswith('delete_fqid_'):
+                    Fastq.objects.filter(
+                    id=int(key.split('delete_fqid_')[1])).delete()
 
 def make_fastq(ip_fastq_path, ip_adapter_path, ip_rep, sminput_fastq_path,
                sminput_adapter_path, sminput_rep,
@@ -80,3 +188,76 @@ def make_fastq(ip_fastq_path, ip_adapter_path, ip_rep, sminput_fastq_path,
             sminput_complete=False
         )
     return
+
+def CLIP_yaml(clip):
+    '''
+    Create a dictionary to generate YAML from CLIP form
+    '''
+    field_dict = dict()
+    field_dict['samples'] = []
+
+    for field, value in clip.__dict__.items():
+        if field == 'fastqs':
+            for i in value.split(','):
+                fastq = Fastq.objects.get(pk=i)
+                field_dict['samples'].append([
+                    {'name': fastq.ip_title,
+                        'read1': {'class': 'File', 'path': fastq.ip_path},
+                        'adapters': {'class': 'File',
+                                     'path': fastq.ip_adapter_path}},
+                    {'name': fastq.sminput_title,
+                        'read1': {'class': 'File', 'path': fastq.sminput_path},
+                        'adapters': {'class': 'File',
+                                     'path': fastq.sminput_adapter_path}},
+                ])
+        elif field in ['speciesGenomeDir', 'repeatElementGenomeDir']:
+            field_dict[field] = {'class': 'Directory', 'path': value}
+        elif field in ['chrom_sizes', 'blacklist_file']:
+            field_dict[field] = {'class': 'File', 'path': value}
+        elif field != '_state' and field != 'barcode_file' and field != 'id':
+            field_dict[field] = value
+    return "#!/usr/bin/env eCLIP_singleend\n" + yaml.dump(
+        field_dict, default_flow_style=False)
+
+def skipper_tsv(clip):
+    '''
+    Create TSV of SKIPPER form
+    '''
+    rows = ','.join([
+        'Experiment', 'Sample', 'Cells',
+        'Input_replicate', 'Input_adapter', 'Input_fastq',
+        'CLIP_replicate', 'CLIP_adapter', 'CLIP_fastq',
+        '\n'
+    ])
+
+    for field, value in clip.__dict__.items():
+        if field == 'fastqs':
+            for i in value.split(','):
+                fastq = Fastq.objects.get(pk=i)
+                row = ','.join([
+                    fastq.experiment,
+                    fastq.sample,
+                    fastq.cells,
+                    str(fastq.ip_rep),
+                    fastq.sminput_adapter_path,
+                    fastq.sminput_path,
+                    str(fastq.sminput_rep),
+                    fastq.ip_adapter_path,
+                    fastq.ip_path,
+                ])
+                rows += row + "\n"
+    return rows
+
+def skipper_config(clip):
+    '''
+    Create CONFIG from SKIPper form
+    '''
+    config_string = ""
+    for field, value in clip.__dict__.items():
+        if field == 'informative_read' or field == 'uninformative_read' or \
+                field == 'umi_size':
+            config_string += f'{field.upper()} = {value}\n'
+        elif field != '_state' and field != 'id' and \
+                field != 'barcode_file' and field != 'fastqs':
+            config_string += f'{field.upper()} = \"{value}\"\n'
+    return config_string
