+++ requirements.txt
@@ -1,20 +1,35 @@
 asgiref==3.5.2
+attrs==22.2.0
+awsebcli==3.20.3
+bcrypt==4.0.1
 beautifulsoup4==4.11.1
+blessed==1.19.1
+botocore==1.29.41
 Brotli==1.0.9
+cached-property==1.5.2
+cement==2.8.2
 certifi==2022.9.24
 cffi==1.15.1
 charset-normalizer==2.1.1
 click==8.1.3
+colorama==0.4.3
 cryptography==36.0.2
 defusedxml==0.7.1
 Django==4.1.1
 django-admin==2.0.1
+django-bootstrap-themes==3.3.6
 django-excel-base==1.0.4
 django-excel-response2==3.0.2
+django-multiselectfield==0.1.12
 django-six==1.0.4
+docker==4.4.4
+docker-compose==1.25.5
+dockerpty==0.4.1
+docopt==0.6.2
 ecdsa==0.18.0
 Flask==2.2.2
 Flask-Compress==1.13
+future==0.16.0
 globus-cli==3.8.0
 globus-sdk==3.11.0
 idna==3.4
@@ -22,17 +37,22 @@ importlib-metadata==4.12.0
 itsdangerous==2.1.2
 Jinja2==3.1.2
 jmespath==0.10.0
+jsonschema==3.2.0
 MarkupSafe==2.1.1
 mysqlclient==2.1.1
 numpy==1.23.3
 oauthlib==3.2.1
 packaging==21.3
 pandas==1.5.0
+paramiko==2.11.0
+pathspec==0.9.0
 plotly==5.10.0
 pyasn1==0.4.8
 pycparser==2.21
 PyJWT==2.5.0
+PyNaCl==1.5.0
 pyparsing==3.0.9
+pyrsistent==0.18.1
 python-dateutil==2.8.2
 python-dotenv==0.20.0
 python-jose==3.3.0
@@ -43,16 +63,20 @@ requests==2.28.1
 requests-oauthlib==1.3.1
 rsa==4.9
 screen==1.0.1
+semantic-version==2.8.5
 six==1.16.0
 social-auth-app-django==5.0.0
 social-auth-core==4.3.0
 soupsieve==2.3.2.post1
 sqlparse==0.4.2
 tenacity==8.1.0
+termcolor==1.1.0
+texttable==1.6.4
 types-cryptography==3.3.23
-typing-extensions==4.3.0
+typing_extensions==4.3.0
 urllib3==1.26.12
+wcwidth==0.1.9
+websocket-client==0.59.0
 Werkzeug==2.2.2
 xlwt==1.3.0
 zipp==3.8.1
-django-bootstrap-themes==3.3.6
